requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9
itsdangerous>=2.1.2
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.2.0